Move zeta2 to its own module within edit_prediction
