+++ crates/cloud_llm_client/src/predict_edits_v3.rs
@@ -77,7 +77,7 @@ pub enum PromptFormat {
 }
 
 impl PromptFormat {
-    pub const DEFAULT: PromptFormat = PromptFormat::OldTextNewText;
+    pub const DEFAULT: PromptFormat = PromptFormat::Minimal;
 }
 
 impl Default for PromptFormat {
